Dependency Injection
Question #8
This is a design pattern that removes the dependency of the programs. In
such case we provide the information from the external source such as
XML file. It makes our code loosely coupled and easier for testing.
See Dependency_injection project that shows an example of wiring CDI
managed beans (MessageServerBean.java) with Facelets page (index.xhtml).